requirements.txt
black>=22.6
cloudinary>=1.40
django-cloudinary-storage>=0.3.0
django-bootstrap-form==3.4
django-crispy-forms>=2.0,<3.0
django-test-without-migrations==0.6
dnspython==1.16.0
graphviz==0.13.2
humanize==0.5.1
inflect==5.3.0
Jinja2==2.11.3
MarkupSafe==2.0.1 # v2.1 removes code used by jinja2 v2.11.3
Pillow>=10.1.0
python-decouple==3.3
requests==2.31.0
Unidecode==1.1.1
django-sendgrid-v5==1.2.1
evennia==3.0.0   # example; use your real version
psycopg[binary]>=3.1
